Remove conditional compilation
