+++ wrap.go
@@ -6,6 +6,22 @@ type Try[A any] struct {
 	Error error
 }
 
+// Stream is a type alias for a channel of [Try] containers.
+// This alias is optional, but it can make the code more readable.
+//
+// Before:
+//
+//	func StreamUsers() <-chan rill.Try[*User] {
+//		...
+//	}
+//
+// After:
+//
+//	func StreamUsers() rill.Stream[*User] {
+//		...
+//	}
+type Stream[T any] = <-chan Try[T]
+
 // Wrap converts a value and/or error into a [Try] container.
 // It's a convenience function to avoid creating a [Try] container manually and benefit from type inference.
 //
